fix wierd bug
