+++ examples/mxnet/gcn/README.md
@@ -29,7 +29,7 @@ Results
 -------
 Run with following (available dataset: "cora", "citeseer", "pubmed")
 ```bash
-DGLBACKEND=mxnet python3 gcn.py --dataset cora --gpu 0
+DGLBACKEND=mxnet python3 train.py --dataset cora --gpu 0
 ```
 
 * cora: ~0.810 (paper: 0.815)
