+++ examples/pytorch/gcn/README.md
@@ -21,13 +21,14 @@ The folder contains three implementations of GCN:
 - `gcn_mp.py` uses user-defined message and reduce functions.
 - `gcn_spmv.py` improves from `gcn_mp.py` by using DGL's builtin functions
    so SPMV optimization could be applied.
+Modify `train.py` to switch between different implementations.
 
 Results
 -------
 
 Run with following (available dataset: "cora", "citeseer", "pubmed")
 ```bash
-python gcn.py --dataset cora --gpu 0
+python train.py --dataset cora --gpu 0
 ```
 
 * cora: ~0.810 (0.79-0.83) (paper: 0.815)
